Refactor: constants 파일 생성 및 상수 정의
